scripts/otel/requirements.txt
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
